pyproject.toml
[tool.poetry]
name = "myproject"
version = "0.1.0"
description = ""
authors = ["minagawa"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.9.13"
selenium = "^4.15.2"
beautifulsoup4 = "^4.12.2"
requests = "^2.31.0"
tqdm = "^4.66.1"
pandas = "^2.1.3"
matplotlib = "^3.8.2"
langchain = "^0.0.348"
faiss-cpu = "^1.7.4"
streamlit = "^1.29.0"
python-dotenv = "^1.0.0"
openai = "^1.3.8"
google-api-python-client = "^2.110.0"
arxiv = "^2.0.0"
scipy = "^1.11.4"
html2text = "^2020.1.16"
tiktoken = "^0.5.2"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
